DAOS-2429 object: predict to fetch from leader replica
If the client has been asked to fetch (list/query) from leader replica,
then means that related data is associated with some prepared DTX that
may be committable on the leader replica. According to our current DTX
batched commit policy, it is quite possible that such DTX is not ready
to be committed, or it is committable but cached on the leader replica
for some time. On the other hand, such DTX may contain more data update
than current fetch. If the subsequent fetch against the same redundancy
group come very soon (within the OBJ_FETCH_LEADER_INTERVAL), then it is
possible that related target for the next fetch is covered by the same
DTX that is still not committed yet. If the assumption is right, asking
the application to fetch from leader replica directly can avoid one RPC
round-trip with non-leader replica. If such assumption is wrong, it may
increase the server load on which the leader replica resides in a short
time but it will not correctness issues.

Signed-off-by: Fan Yong <fan.yong@intel.com>